Make AalCameraService advance status as things happen
Previously, the camera's status is never updated because we assume that
state change is synchronous. However, we've learned from other plugins
that it isn't always the case. Applications need to use the status
property to find out if state change is complete or not.

This commit make sure that the status property reflects the actual
status of the camera. This is done by letting AalCameraService tells
AalCameraControl when the service connect/disconnect the camera, start/
stop preview, and when the capture mode is changed. This should cover
all changes in status required.