AudioCapture: Add 1 sec timeout for writing audio
On the Pixel 3a a race condition can happen when recording
video with audio, which we almost always lose.
This is manifested in the Halium-side RecordThread stopping
whereas the Ubuntu Touch side doesn't know what to do
and hangs at the blocking write to the micshm fifo.

To counter this add a 1 second timeout to the write
to get out of the hanging state and have it exit cleanly.

This fixes hangs when stopping video recording on at
least the Pixel 3a (sargo), and potentially others.